+++ R/shims.r
@@ -3,7 +3,10 @@
 # @param pkg A path or package object
 insert_imports_shims <- function(pkg = ".") {
   pkg <- as.package(pkg)
-  assign("system.file", shim_system.file, pos = imports_env(pkg))
+
+  imp_env <- imports_env(pkg)
+  imp_env$system.file <- shim_system.file
+  imp_env$library.dynam.unload <- shim_library.dynam.unload
 }
 
 # Create a new environment as the parent of global, with devtools versions of
@@ -90,4 +93,23 @@ shim_system.file <- function(..., package = "base", lib.loc = NULL,
     # be installed. To fully duplicate R's package-building and installation
     # behavior would be complicated, so we'll just use this simple method.
   }
+}
+
+shim_library.dynam.unload <- function(chname, libpath,
+                                      verbose = getOption("verbose"),
+                                      file.ext = .Platform$dynlib.ext) {
+
+  # If package was loaded by devtools, we need to unload the dll ourselves
+  # because libpath works differently from installed packages.
+  if (!is.null(dev_meta(chname))) {
+    try({
+      pkg <- as.package(libpath)
+      unload_dll(pkg)
+    })
+    return()
+  }
+
+  # Should only reach this in the rare case that the devtools-loaded package is
+  # trying to unload a different package's DLL.
+  library.dynam.unload(chname, libpath, verbose, file.ext)
 }